refactor: menu to components
